chore: bump deps and move tests
