+++ packages/form-core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tanstack/form-core",
-  "version": "0.42.1",
+  "version": "0.42.0",
   "description": "Powerful, type-safe, framework agnostic forms.",
   "author": "tannerlinsley",
   "license": "MIT",
